favoriteRepositoryProviderの変数名を変更とファイル名を変更
